Module Sidekick_base.Base_types

Basic type definitions for Sidekick_base

module Vec = Sidekick_util.Vec

module Log = Sidekick_util.Log

module Fmt = CCFormat

module CC_view = Sidekick_core.CC_view

module Proof_ser = Sidekick_base_proof_trace.Proof_ser

module Storage = Sidekick_base_proof_trace.Storage

val hash_z : Z.t -> int

val hash_q : Q.t -> CCHash.hash

module LRA_pred : sig ... end

module LRA_op : sig ... end

module LRA_view : sig ... end

module LIA_pred = LRA_pred

module LIA_op = LRA_op

module LIA_view : sig ... end

type term = {

mutable term_id : int;

mutable term_ty : ty;

term_view : term term_view;

}

Term.

A term, with its own view, type, and a unique identifier. Do not create directly, see Term.

and 'a term_view =

| Bool of bool

| App_fun of fun_ * 'a Sidekick_util.IArray.t

| Eq of 'a * 'a

| Not of 'a

| Ite of 'a * 'a * 'a

| LRA of 'a LRA_view.t

| LIA of 'a LIA_view.t

Shallow structure of a term.

A term is a DAG (direct acyclic graph) of nodes, each of which has a term view.

and fun_ = {

fun_id : ID.t;

fun_view : fun_view;

}

type of function symbols

and fun_view =

| Fun_undef of fun_ty

| Fun_select of select

| Fun_cstor of cstor

| Fun_is_a of cstor

| Fun_def of {

pp : 'a. ( 'a Fmt.printer -> 'a Sidekick_util.IArray.t Fmt.printer ) option;

abs : self:term -> term Sidekick_util.IArray.t -> term * bool;

do_cc : bool;

relevant : 'a. ID.t -> 'a Sidekick_util.IArray.t -> int -> bool;

ty : ID.t -> term Sidekick_util.IArray.t -> ty;

eval : value Sidekick_util.IArray.t -> value;

}	(*
Methods on the custom term view whose arguments are 'a. Terms must be printable, and provide some additional theory handles.

relevant must return a subset of args (possibly the same set). The terms it returns will be activated and evaluated whenever possible. Terms in args \ relevant args are considered for congruence but not for evaluation.

*)

and fun_ty = {

fun_ty_args : ty list;

fun_ty_ret : ty;

}

Function type

and ty = {

mutable ty_id : int;

ty_view : ty_view;

}

Hashconsed type

and ty_view =

| Ty_bool

| Ty_real

| Ty_int

| Ty_atomic of {

def : ty_def;

args : ty list;

mutable finite : bool;

}

and ty_def =

| Ty_uninterpreted of ID.t

| Ty_data of {

data : data;

}

| Ty_def of {

id : ID.t;

pp : ty Fmt.printer -> ty list Fmt.printer;

default_val : value list -> value;

}

and data = {

data_id : ID.t;

data_cstors : cstor Sidekick_base__.ID.Map.t lazy_t;

data_as_ty : ty lazy_t;

}

and cstor = {

cstor_id : ID.t;

cstor_is_a : ID.t;

mutable cstor_arity : int;

cstor_args : select list lazy_t;

cstor_ty_as_data : data;

cstor_ty : ty lazy_t;

}

and select = {

select_id : ID.t;

select_cstor : cstor;

select_ty : ty lazy_t;

select_i : int;

}

and value =

| V_bool of bool

| V_element of {

id : ID.t;

ty : ty;

}	(*
a named constant, distinct from any other constant
*)

| V_cstor of {

c : cstor;

args : value list;

}

| V_custom of {

view : value_custom_view;

pp : value_custom_view Fmt.printer;

eq : value_custom_view -> value_custom_view -> bool;

hash : value_custom_view -> int;

}	(*
Custom value
*)

| V_real of Q.t

Semantic values, used for models (and possibly model-constructing calculi)

and value_custom_view = ..

type definition = ID.t * ty * term

type statement =

| Stmt_set_logic of string

| Stmt_set_option of string list

| Stmt_set_info of string * string

| Stmt_data of data list

| Stmt_ty_decl of ID.t * int

| Stmt_decl of ID.t * ty list * ty

| Stmt_define of definition list

| Stmt_assert of term

| Stmt_assert_clause of term list

| Stmt_check_sat of (bool * term) list

| Stmt_get_model

| Stmt_get_value of term list

| Stmt_exit

val term_equal_ : term -> term -> bool

val term_hash_ : term -> int

val term_cmp_ : term -> term -> int

val fun_compare : fun_ -> fun_ -> int

val pp_fun : CCFormat.t -> fun_ -> unit

val id_of_fun : fun_ -> ID.t

val eq_ty : ty -> ty -> bool

val eq_cstor : cstor -> cstor -> bool

val eq_value : value -> value -> bool

val hash_value : value Sidekick_util.Hash.t

val pp_value : value Sidekick_util.Util.printer

val pp_db : Stdlib.Format.formatter -> (int * 'a) -> unit

val pp_ty : ty Sidekick_util.Util.printer

val pp_term_view_gen : pp_id:( Fmt.t -> ID.t -> unit ) -> pp_t:'a Fmt.printer -> Fmt.t -> 'a term_view -> unit

val pp_term_top : ids:bool -> Fmt.t -> term -> unit

val pp_term : Fmt.t -> term -> unit

val pp_term_view : Fmt.t -> term term_view -> unit

module Ty : sig ... end

Types

module Fun : sig ... end

Function symbols

module Term_cell : sig ... end

module Term : sig ... end

Term creation and manipulation

module Value : sig ... end

Values (used in models)

module Data : sig ... end

Datatypes

module Select : sig ... end

Datatype selectors.

module Cstor : sig ... end

Datatype constructors.

module Statement : sig ... end

Statements.