Module Base_types.Cstor

Datatype constructors.

A datatype has one or more constructors, each of which is a special kind of function symbol. Constructors are injective and pairwise distinct.

type t = cstor = {

cstor_id : ID.t;

cstor_is_a : ID.t;

mutable cstor_arity : int;

cstor_args : select list lazy_t;

cstor_ty_as_data : data;

cstor_ty : ty lazy_t;

}

val id : t -> ID.t

val ty_args : t -> ty Iter.t

val equal : cstor -> cstor -> bool

val pp : CCFormat.t -> t -> unit