Module Base_types.Select

Datatype selectors.

A selector is a kind of function that allows to obtain an argument of a given constructor.

type t = select = {

select_id : ID.t;

select_cstor : cstor;

select_ty : ty lazy_t;

select_i : int;

}

val ty : t -> ty