Module Base_types.Term

Term creation and manipulation

type t = term = {

mutable term_id : int;

mutable term_ty : ty;

term_view : t term_view;

}

type 'a view = 'a term_view =

| Bool of bool

| App_fun of fun_ * 'a Sidekick_util.IArray.t

| Eq of 'a * 'a

| Not of 'a

| Ite of 'a * 'a * 'a

| LRA of (Q.t, 'a) lra_view

val id : t -> int

val view : t -> term view

val ty : t -> Ty.t

val equal : t -> t -> bool

val compare : t -> t -> int

val hash : t -> int

type store

val create : ?size:int -> unit -> store

val make : store -> t view -> t

val true_ : store -> t

val false_ : store -> t

val bool : store -> bool -> t

val const : store -> fun_ -> t

val app_fun : store -> fun_ -> t Sidekick_util.IArray.t -> t

val app_fun_l : store -> fun_ -> t list -> t

val eq : store -> t -> t -> t

val not_ : store -> t -> t

val ite : store -> t -> t -> t -> t

val app_undefined : store -> ID.t -> Ty.Fun.t -> t Sidekick_util.IArray.t -> t

app_undefined store f ty args is app store (Fun.mk_undef f ty) args. It builds a function symbol and applies it into a term immediately

val const_undefined : store -> ID.t -> Ty.t -> t

const_undefined store f ty is const store (Fun.mk_undef_const f ty). It builds a constant function symbol and makes it into a term immediately.

val select : store -> select -> t -> t

val app_cstor : store -> cstor -> t Sidekick_util.IArray.t -> t

val is_a : store -> cstor -> t -> t

val lra : store -> (Q.t, t) lra_view -> t

module LRA : sig ... end

Helpers for LRA

val abs : store -> t -> t * bool

Obtain unsigned version of t, + the sign as a boolean

module Iter_dag : sig ... end

val iter_dag_with : order:Iter_dag.order -> t -> t Iter.t

val iter_dag : t -> t Iter.t

val map_shallow : store -> (t -> t) -> t -> t

val pp : t Fmt.printer

Views

val is_true : t -> bool

val is_false : t -> bool

val is_const : t -> bool

val cc_view : t -> (fun_, t, t Iter.t) CC_view.t

val as_fun_undef : t -> (fun_ * Ty.Fun.t) option

val as_bool : t -> bool option

Store

val store_size : store -> int

val store_iter : store -> term Iter.t

Containers

module Tbl : CCHashtbl.S with type key = t

module Map : CCMap.S with type key = t

module Set : CCSet.S with type elt = t