Module Sidekick_base.Form

Formulas (boolean terms).

This module defines function symbols, constants, and views to manipulate boolean formulas in Sidekick_base. This is useful to have the ability to use boolean connectives instead of being limited to clauses; by using Sidekick_th_bool_static, the formulas are turned into clauses automatically for you.

type term = Types_.Term.t

type 'a view = 'a Sidekick_core.Bool_view.t =

| B_bool of bool

| B_not of 'a

| B_and of 'a list

| B_or of 'a list

| B_imply of 'a * 'a

| B_equiv of 'a * 'a

| B_xor of 'a * 'a

| B_eq of 'a * 'a

| B_neq of 'a * 'a

| B_ite of 'a * 'a * 'a

| B_atom of 'a

val view : term -> term view

val bool : Types_.Term.store -> bool -> term

val not_ : Types_.Term.store -> term -> term

val and_ : Types_.Term.store -> term -> term -> term

val or_ : Types_.Term.store -> term -> term -> term

val eq : Types_.Term.store -> term -> term -> term

val neq : Types_.Term.store -> term -> term -> term

val imply : Types_.Term.store -> term -> term -> term

val equiv : Types_.Term.store -> term -> term -> term

val xor : Types_.Term.store -> term -> term -> term

val ite : Types_.Term.store -> term -> term -> term -> term

val distinct_l : Types_.Term.store -> term list -> term

val and_l : Types_.Term.store -> term list -> term

val or_l : Types_.Term.store -> term list -> term

val imply_l : Types_.Term.store -> term list -> term -> term

val mk_of_view : Types_.Term.store -> term view -> term