Module Sidekick_base.Solver_arg

Concrete implementation of Sidekick_core.TERM

this module gathers most definitions above in a form that is compatible with what Sidekick expects for terms, functions, etc.

include Sidekick_core.TERM with type Term.t = Base_types.Term.t and type Fun.t = Base_types.Fun.t and type Ty.t = Base_types.Ty.t and type Term.store = Base_types.Term.store and type Ty.store = Base_types.Ty.store

module Fun : sig ... end

A function symbol, like "f" or "plus" or "is_human" or "socrates"

module Ty : sig ... end

Types

module Term : sig ... end

Term structure.