Module Sidekick_base.Statement

Statements.

A statement is an instruction for the SMT solver to do something, like asserting that a formula is true, declaring a new constant, or checking satisfiabilty of the current set of assertions.

type t = Types_.statement =

| Stmt_set_logic of string

| Stmt_set_option of string list

| Stmt_set_info of string * string

| Stmt_data of Types_.data list

| Stmt_ty_decl of {

name : ID.t;

arity : int;

ty_const : Types_.ty;

}
(*
new atomic cstor
*)

| Stmt_decl of {

name : ID.t;

ty_args : Types_.ty list;

ty_ret : Types_.ty;

const : Types_.term;

}

| Stmt_define of Types_.definition list

| Stmt_assert of Types_.term

| Stmt_assert_clause of Types_.term list

| Stmt_check_sat of (bool * Types_.term) list

| Stmt_get_model

| Stmt_get_value of Types_.term list

| Stmt_exit

include Sidekick_sigs.PRINT with type t := t

val pp : t Sidekick_sigs.printer