Module Sidekick_base__Model

Models

A model is a solution to the satisfiability question, created by the SMT solver when it proves the formula to be satisfiable.

A model gives a value to each term of the original formula(s), in such a way that the formula(s) is true when the term is replaced by its value.

module Val_map : sig ... end

module Fun_interpretation : sig ... end

Model for function symbols.

type t = {

values : Sidekick_base.Base_types.Value.t Sidekick_base__.Base_types.Term.Map.t;

funs : Fun_interpretation.t Sidekick_base__.Base_types.Fun.Map.t;

}

Model

val empty : t

Empty model

val add : Sidekick_base.Base_types.Term.t -> Sidekick_base.Base_types.Value.t -> t -> t

val mem : Sidekick_base.Base_types.Term.t -> t -> bool

val find : Sidekick_base.Base_types.Term.t -> t -> Sidekick_base.Base_types.Value.t option

val merge : t -> t -> t

val pp : t CCFormat.printer

val eval : t -> Sidekick_base.Base_types.Term.t -> Sidekick_base.Base_types.Value.t option

eval m t tries to evaluate term t in the model. If it succeeds, the value is returned, otherwise None is.