Module Sidekick_base_proof_trace

Proof trace with serialization

This library is useful to serialize a series of reasoning steps in memory or into a file, to be able to reconstruct a proper proof later.

module Proof_ser : sig ... end

module Storage : sig ... end

val iter_steps_backward : Storage.t -> Proof_ser.Step.t Iter.t