Module Sidekick_base.Form

Formulas (boolean terms).

This module defines function symbols, constants, and views to manipulate boolean formulas in Sidekick_base. This is useful to have the ability to use boolean connectives instead of being limited to clauses; by using Sidekick_th_bool_static, the formulas are turned into clauses automatically for you.

module T = Base_types.Term

module Ty = Base_types.Ty

module Fun = Base_types.Fun

module Value = Base_types.Value

exception Not_a_th_term

val id_and : ID.t

val id_or : ID.t

val id_imply : ID.t

val view_id : ID.t -> 'a Sidekick_util.IArray.t -> ( 'a, 'a Sidekick_util.IArray.iter ) Sidekick_th_bool_static.bool_view

val view_as_bool : T.t -> ( T.t, T.t Sidekick_util.IArray.iter ) Sidekick_th_bool_static.bool_view

module Funs : sig ... end

val as_id : ID.t -> T.t -> T.t Sidekick_util.IArray.t option

val flatten_id : ID.t -> bool -> T.t list -> T.t list

val and_l : T.store -> T.t list -> T.t

val or_l : T.store -> T.t list -> T.t

val and_ : T.store -> T.t -> T.t -> T.t

val or_ : T.store -> T.t -> T.t -> T.t

val and_a : T.store -> T.t Sidekick_util.IArray.t -> T.t

val or_a : T.store -> T.t Sidekick_util.IArray.t -> T.t

val eq : T.store -> T.t -> T.t -> T.t

val not_ : T.store -> T.t -> T.t

val ite : T.store -> T.t -> T.t -> T.t -> T.t

val equiv : T.store -> T.t -> T.t -> T.t

val neq : T.store -> T.t -> T.t -> T.t

val imply_a : T.store -> T.t Sidekick_util.IArray.t -> T.t -> T.t

val imply_l : T.store -> T.t list -> T.t -> T.t

val imply : T.store -> T.t -> T.t -> T.t

val xor : T.store -> T.t -> T.t -> T.t

val distinct_l : T.store -> T.t CCList.t -> T.t

val mk_bool : T.store -> ( T.t, T.t Sidekick_util.IArray.t ) Sidekick_th_bool_static.bool_view -> T.t

module Gensym : sig ... end